changes pushed

+++ src/app/components/client/client.component.ts
@@ -1,5 +1,5 @@
 import { Component, inject, OnInit } from '@angular/core';
-import { KENDO_GRID } from '@progress/kendo-angular-grid';
+import { EditEvent, KENDO_GRID } from '@progress/kendo-angular-grid';
 import { IClient } from '../../model/interface/client';
 import { ClientService } from '../../services/client.service';
 import { APIResponseModel } from '../../model/interface/role';
@@ -61,6 +61,7 @@ export class ClientComponent implements OnInit{
       //   console.log(res.data,"res.data")
       // }
       console.log(res.data,"res.data")
+      this.clientObj = new Client();
     })
   }
 
@@ -88,73 +89,22 @@ export class ClientComponent implements OnInit{
     })
   }
 
-  // onContextMenuSelect(event: any) {
-  //   const selectedItem = event.item.text;
-  //   console.log(selectedItem,"selectedItem ")
-  //   if(selectedItem == "Add Client") {
-  //     this.showClientDialog = true
-  //     console.log("add client selectd ")
-  //     console.log(selectedItem,"add selectedItem ")
-  //   }else if(selectedItem == "Edit Client"){
-
-
-  //   }else if(selectedItem == "Delete Client"){
-  //     this.deleteClientDialog = true;
-  //   }
-  // }
-
-  onContextMenuSelect(event: any): void {
-    const action = event.item.text;
-  
-    switch (action) {
-      case 'Delete Client':
-        if (this.selectedClient) {
-          this.deleteClientDialog = true; // Show confirmation
-        } else {
-          alert('Please select a client to delete.');
-        }
-        break;
-  
-      case 'Edit Client':
-        if (this.selectedClient) {
-          this.isNew = false;
-          this.showClientDialog = true;
-        } else {
-          alert('Please select a client to edit.');
-        }
-        break;
-  
-      case 'Add Client':
-        this.isNew = true;
-        // this.clientObj = {};
-        this.showClientDialog = true;
-        break;
-  
-      default:
-        break;
-    }
-  }
-
-  
-
-  onGridSelectionChange(event: any): void {
-    this.selectedClient = event.selectedRow[0].dataItem;
-    console.log(this.selectedClient,"this.selectedClient");
-    this.clientObj = {...this.selectedClient};
-    console.log(this.clientObj ,"this.clientObj ")
-
-  }
-
-  onClose() {
+   onClose() {
     this.showClientDialog = false;
     this.deleteClientDialog = false;
+    this.clientObj = new Client();
   }
 
   onSubmit() {
     console.log("submit pressed")
     this.clientService.addUpdateClient(this.clientObj).subscribe((res:APIResponseModel)=>{
       if(res.result){
-        alert("Client Created  SuccessFully!!");
+        if(this.isNew){
+          alert("Client Created  SuccessFully!!");
+        }else{
+          alert("Client Updated  SuccessFully!!");
+        }
+        
         this.showClientDialog = false
         this.getAllClient();
 
@@ -167,10 +117,85 @@ export class ClientComponent implements OnInit{
     })
   }
 
+  onContextMenuSelect(event: any) {
+    const selectedItem = event.item.text;
+    console.log(selectedItem,"selectedItem ")
+    if(selectedItem == "Add Client") {
+      this.isNew = true
+      this.showClientDialog = true
+      console.log("add client selectd ")
+      console.log(selectedItem,"add selectedItem ")
+    }else if(selectedItem == "Edit Client"){
+      this.editClient(event);
+      // this.isNew = false;
+      // this.showClientDialog = true;
+    }else if(selectedItem == "Delete Client"){
+      this.deleteClientDialog = true;
+      if(this.selectedClient && this.selectedClient.clientId){
+        this.onDeleteClient(this.selectedClient.clientId);
+      }else {
+        alert("No ClientId is selected")
+      }
+    }
+  }
+
+
+  onGridSelectionChange(event:any): void {
+    this.selectedClient = event.selectedRow[0].dataItem;
+    console.log(this.selectedClient,"this.selectedClient")
+  }
+
+  onClickForContextMenu(event: MouseEvent){
+    const targetRow = (event.target as HTMLElement).closest('kendo-grid-list tr');
+    console.log(targetRow,"targetRow");
+
+    if(targetRow){
+      const rowIndex = Array.from(targetRow.parentElement!.children).indexOf(targetRow);
+      console.log(rowIndex,"rowIndex")
+
+      this.selectedClient = this.clientList[rowIndex];
+      console.log(this.selectedClient,"this.selectedClient")
+
+       const objet = this.selectedClient.clientId;
+      console.log(objet,"objet")
+
+    }
+  }
+
+
+
   onDeleteClient(clientId: number){
-    this.clientService.deleteClient(clientId).subscribe((res: APIResponseModel)=> {
-    console.log(res.message,"res.message")
+    const isDelete = confirm("ARE YOU SURE YOU WANT TO DELETE?")
+    console.log(clientId,"clientId")
+    if(isDelete){
+       this.clientService.deleteClient(clientId).subscribe((res: APIResponseModel)=> {
+      console.log(res,"res")
+      if(res.result){
+        alert("Deleted Successfully!")
+        this.getAllClient();
+      }else {
+        alert(res.message + "Delete Failed" )
+        this.getAllClient();
+      }
     })
+    } else {
+      this.deleteClientDialog = false;
+    }
+  }
+
+
+  editClient(event: EditEvent) {
+    this.isNew = false;
+    const clientId = this.selectedClient.clientId;
+     this.clientService.getClientById(clientId).subscribe((res:APIResponseModel)=>{
+      this.clientObj = res.data
+      this.showClientDialog = true;
+    })
+  }
+
+  resetFilter() {
+    this.clientObj.clientId = 0;
+    this.getAllClient();
   }
